Moved raster related parts of VtkMeshConverter to seperate file and fixed all calls
