+++ Tests/MeshLib/TestNodeSearch.cpp
@@ -14,8 +14,7 @@
 #include "MeshLib/Node.h"
 #include "MeshLib/Elements/Element.h"
 #include "MeshLib/MeshSearch/NodeSearch.h"
-#include "MeshLib/MeshGenerators/VtkMeshConverter.h"
-//#include "MeshLib/MeshGenerators/ConvertRasterToMesh.h"
+#include "MeshLib/MeshGenerators/RasterToMesh.h"
 #include "MeshLib/MeshEditing/DuplicateMeshComponents.h"
 
 #include "GeoLib/Raster.h"
@@ -26,9 +25,7 @@ TEST(NodeSearch, UnusedNodes)
 	GeoLib::RasterHeader const header = 
 		{4,3,MathLib::Point3d(std::array<double,3>{{0,0,0}}),1,-9999};
 	GeoLib::Raster const raster(header,pix.begin(), pix.end());
-	MeshLib::Mesh* mesh (MeshLib::VtkMeshConverter::convertRasterToMesh(raster, MeshLib::MeshElemType::TRIANGLE, MeshLib::UseIntensityAs::ELEVATION));
-	//MeshLib::ConvertRasterToMesh conv(raster, MeshLib::MeshElemType::TRIANGLE, MeshLib::UseIntensityAs::ELEVATION);
-	//MeshLib::Mesh* mesh = conv.execute();
+	MeshLib::Mesh* mesh (MeshLib::RasterToMesh::convert(raster, MeshLib::MeshElemType::TRIANGLE, MeshLib::UseIntensityAs::ELEVATION));
 	MeshLib::NodeSearch ns(*mesh);
 	ns.searchUnused();
 	std::vector<std::size_t> u_nodes = ns.getSearchedNodeIDs();
